Implement a nicer cli with argp
With the shiny new cli, you can:
- continue to connect to a server
- continue to load from a file or stdin
- specify a height/width for the canvas, (which will also resize it
  after loading a file)
- get an automagic `git describe` version
- file an issue on github
- get an introduction to the interface

In doing this, I refactored the state initialization into a function
and added an arguments_t struct that represents the state of the
arguments passed in.

I also made some network-related variables global - not sure if that
is the best way to do this.