Add version negotiation to network protocol
Before making more breaking changes to the networking protocol (like
sending user coordinates back and forth for cursor support), it seemed
best to add a method of negotiating a protocol version so things can
fail or downgrade more gracefully.

As implemented here, previous versions of servers and clients will not
work with these updated versions. This would be a pretty big problem,
except that as far as I know nobody uses this project regularly,
especially in a networked/remote capacity.

Because previously the client expected canvas data to be sent over
immediately, the only way that I can think of to support older clients
while not requiring that canvas send is to add a timeout on the server
that waits for a version number and then sends the canvas if nothing is
sent, which seems pretty unideal.